Move pkg/progress to api/pkg/progress
