+++ daemon/builder/dockerfile/copy.go
@@ -19,10 +19,10 @@ import (
 	"github.com/docker/docker/daemon/builder/remotecontext/urlutil"
 	"github.com/docker/docker/daemon/internal/system"
 	"github.com/docker/docker/pkg/longpath"
-	"github.com/docker/docker/pkg/progress"
 	"github.com/docker/docker/pkg/streamformatter"
 	"github.com/moby/buildkit/frontend/dockerfile/instructions"
 	"github.com/moby/go-archive"
+	"github.com/moby/moby/api/pkg/progress"
 	"github.com/moby/sys/symlink"
 	"github.com/moby/sys/user"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
