+++ daemon/containerd/image_push.go
@@ -19,8 +19,8 @@ import (
 	"github.com/distribution/reference"
 	"github.com/docker/docker/daemon/internal/metrics"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/pkg/progress"
 	"github.com/docker/docker/pkg/streamformatter"
+	"github.com/moby/moby/api/pkg/progress"
 	"github.com/moby/moby/api/types/auxprogress"
 	"github.com/moby/moby/api/types/events"
 	"github.com/moby/moby/api/types/registry"
