+++ daemon/images/image_push.go
@@ -11,7 +11,7 @@ import (
 	progressutils "github.com/docker/docker/daemon/internal/distribution/utils"
 	"github.com/docker/docker/daemon/internal/metrics"
 	"github.com/docker/docker/daemon/server/backend"
-	"github.com/docker/docker/pkg/progress"
+	"github.com/moby/moby/api/pkg/progress"
 	"github.com/moby/moby/api/types/registry"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 )
