+++ daemon/internal/distribution/config.go
@@ -15,7 +15,7 @@ import (
 	"github.com/docker/docker/daemon/internal/layer"
 	refstore "github.com/docker/docker/daemon/internal/refstore"
 	registrypkg "github.com/docker/docker/daemon/pkg/registry"
-	"github.com/docker/docker/pkg/progress"
+	"github.com/moby/moby/api/pkg/progress"
 	"github.com/moby/moby/api/types/events"
 	"github.com/moby/moby/api/types/registry"
 	"github.com/opencontainers/go-digest"
