+++ daemon/internal/distribution/push.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/containerd/log"
 	"github.com/distribution/reference"
-	"github.com/docker/docker/pkg/progress"
+	"github.com/moby/moby/api/pkg/progress"
 	"github.com/moby/moby/api/types/events"
 )
