+++ daemon/internal/distribution/push_v2_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/docker/docker/daemon/internal/layer"
 	refstore "github.com/docker/docker/daemon/internal/refstore"
 	registrypkg "github.com/docker/docker/daemon/pkg/registry"
-	"github.com/docker/docker/pkg/progress"
+	"github.com/moby/moby/api/pkg/progress"
 	"github.com/moby/moby/api/types/registry"
 	"github.com/opencontainers/go-digest"
 	"gotest.tools/v3/assert"
