+++ daemon/internal/distribution/xfer/download.go
@@ -12,8 +12,8 @@ import (
 	"github.com/docker/docker/daemon/internal/image"
 	"github.com/docker/docker/daemon/internal/layer"
 	"github.com/docker/docker/pkg/ioutils"
-	"github.com/docker/docker/pkg/progress"
 	"github.com/moby/go-archive/compression"
+	"github.com/moby/moby/api/pkg/progress"
 )
 
 const maxDownloadAttempts = 5
