+++ daemon/internal/distribution/xfer/download_test.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/docker/distribution"
 	"github.com/docker/docker/daemon/internal/layer"
-	"github.com/docker/docker/pkg/progress"
+	"github.com/moby/moby/api/pkg/progress"
 	"github.com/opencontainers/go-digest"
 	"github.com/opencontainers/image-spec/identity"
 	"gotest.tools/v3/assert"
