+++ daemon/internal/distribution/xfer/upload.go
@@ -8,7 +8,7 @@ import (
 	"github.com/containerd/log"
 	"github.com/docker/distribution"
 	"github.com/docker/docker/daemon/internal/layer"
-	"github.com/docker/docker/pkg/progress"
+	"github.com/moby/moby/api/pkg/progress"
 )
 
 const maxUploadAttempts = 5
