+++ daemon/internal/image/tarexport/load.go
@@ -19,10 +19,10 @@ import (
 	"github.com/docker/docker/daemon/internal/ioutils"
 	"github.com/docker/docker/daemon/internal/layer"
 	"github.com/docker/docker/daemon/internal/stringid"
-	"github.com/docker/docker/pkg/progress"
 	"github.com/docker/docker/pkg/streamformatter"
 	"github.com/moby/go-archive/chrootarchive"
 	"github.com/moby/go-archive/compression"
+	"github.com/moby/moby/api/pkg/progress"
 	"github.com/moby/moby/api/types/events"
 	"github.com/moby/sys/sequential"
 	"github.com/moby/sys/symlink"
