+++ daemon/pkg/plugin/backend_linux.go
@@ -29,8 +29,8 @@ import (
 	"github.com/docker/docker/errdefs"
 	"github.com/docker/docker/pkg/authorization"
 	"github.com/docker/docker/pkg/pools"
-	"github.com/docker/docker/pkg/progress"
 	"github.com/moby/go-archive/chrootarchive"
+	"github.com/moby/moby/api/pkg/progress"
 	"github.com/moby/moby/api/types"
 	"github.com/moby/moby/api/types/events"
 	"github.com/moby/moby/api/types/filters"
